refactor to extract renotifySDES to service
